requirements.txt
django
gunicorn
django-heroku
requests
psycopg2
dj-database-url
psycopg2-binary
pytest
pytest-django
pytest-html
more-itertools
ipython
attr
tabulate
pandas
django-service-objects
djangorestframework
djangorestframework-jwt
mock
